fix: handle non-list response_stream in HttpResponse.json property
When using the aiohttp backend and Google API returns a 500 error,
the response_stream can be an aiohttp.ClientResponse object instead
of a list. This causes 'TypeError: ClientResponse object is not
subscriptable' when trying to access response_stream[0].

This fix adds a type check before attempting list access, returning
None for non-list response_stream values to allow proper error
handling to continue.

Fixes: https://github.com/googleapis/python-genai/issues/1897